+++ tests/Fingerprint/FingerprintTaskWithAFingerprintAndForceTest.php
@@ -4,11 +4,11 @@ namespace Castor\Tests\Fingerprint;
 
 use Castor\Tests\TaskTestCase;
 
-class FingerprintTaskWithSomeFingerprintWithForceOptionTest extends TaskTestCase
+class FingerprintTaskWithAFingerprintAndForceTest extends TaskTestCase
 {
     use FingerprintedTest;
 
-    // fingerprint:task-with-some-fingerprint-and-force
+    // fingerprint:task-with-a-fingerprint-and-force
     public function test(): void
     {
         $filepath = \dirname(__DIR__, 2) . '/examples/fingerprint_file.fingerprint_single';
@@ -19,9 +19,9 @@ class FingerprintTaskWithSomeFingerprintWithForceOptionTest extends TaskTestCase
 
         file_put_contents($filepath, 'Hello');
 
-        $processFirstRun = $this->runTask(['fingerprint:task-with-some-fingerprint-and-force']);
-        $processSecondRun = $this->runTask(['fingerprint:task-with-some-fingerprint-and-force', '--force']);
-        $processThirdRun = $this->runTask(['fingerprint:task-with-some-fingerprint-and-force']);
+        $processFirstRun = $this->runTask(['fingerprint:task-with-a-fingerprint-and-force']);
+        $processSecondRun = $this->runTask(['fingerprint:task-with-a-fingerprint-and-force', '--force']);
+        $processThirdRun = $this->runTask(['fingerprint:task-with-a-fingerprint-and-force']);
 
         $this->assertStringEqualsFile(__FILE__ . '.output_runnable.txt', $processFirstRun->getOutput());
         $this->assertStringEqualsFile(__FILE__ . '.output_runnable.txt', $processSecondRun->getOutput());
@@ -29,7 +29,7 @@ class FingerprintTaskWithSomeFingerprintWithForceOptionTest extends TaskTestCase
 
         file_put_contents($filepath, 'Hello World');
         // If we don't force, it should re-run the task
-        $processFourthRun = $this->runTask(['fingerprint:task-with-some-fingerprint-and-force']);
+        $processFourthRun = $this->runTask(['fingerprint:task-with-a-fingerprint-and-force']);
         $this->assertStringEqualsFile(__FILE__ . '.output_runnable.txt', $processFourthRun->getOutput());
 
         foreach ([$processFirstRun, $processSecondRun, $processThirdRun, $processFourthRun] as $process) {